Make ConsoleAction private to console player
