+++ royals_core/src/console_player.rs
@@ -6,7 +6,7 @@ use std::{
 use itertools::Itertools;
 
 use crate::{
-    action::{Action, ConsoleAction},
+    action::Action,
     card::Card,
     event::Event,
     play::Play,
@@ -23,12 +23,68 @@ When the card is played an action might be performed based on the type of card i
 At the beginning a card is put to the side, that is hidden an not used except for the special case, when the last card played is a Prince.
 If all opponents are protected one may choose to not do anything.";
 
+#[derive(Debug, PartialEq)]
+enum ConsoleAction {
+    Quit,
+    Rules,
+    CardEffects,
+    Card(Card),
+    Player(PlayerId),
+}
+
+#[derive(Debug, PartialEq, Eq)]
+struct ParseActionError;
+
+impl ConsoleAction {
+    fn info(&self, players: &[Player]) -> String {
+        match self {
+            ConsoleAction::Quit => "quit".to_string(),
+            ConsoleAction::Rules => "display rules".to_string(),
+            ConsoleAction::CardEffects => "display card effects".to_string(),
+            ConsoleAction::Card(c) => c.rule().to_string(),
+            ConsoleAction::Player(id) => players[*id].name.clone(),
+        }
+    }
+
+    fn cmd_str(&self) -> String {
+        match self {
+            ConsoleAction::Quit => "q".to_string(),
+            ConsoleAction::Rules => "r".to_string(),
+            ConsoleAction::CardEffects => "c".to_string(),
+            ConsoleAction::Card(c) => c.to_string(),
+            ConsoleAction::Player(id) => "".to_string() + &id.to_string(),
+        }
+    }
+}
+
+impl FromStr for ConsoleAction {
+    type Err = ParseActionError;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "q" => Ok(ConsoleAction::Quit),
+            "r" => Ok(ConsoleAction::Rules),
+            "c" => Ok(ConsoleAction::CardEffects),
+            _ => {
+                if let Ok(p) = Card::from_str(s) {
+                    Ok(ConsoleAction::Card(p))
+                } else {
+                    if let Ok(p) = usize::from_str(s) {
+                        Ok(ConsoleAction::Player(p))
+                    } else {
+                        Err(ParseActionError)
+                    }
+                }
+            }
+        }
+    }
+}
+
 pub struct ConsolePlayer {
     pub id: PlayerId,
 }
 
 impl ConsolePlayer {
-    pub fn query_user(
+    fn query_user(
         &self,
         cmds: Vec<ConsoleAction>,
         prompt: &str,
