+++ royals_core/src/lib.rs
@@ -1,7 +1,6 @@
 use event::Event;
 use game_state::GameState;
 use play::Play;
-use player::PlayerId;
 
 mod action;
 mod card;
